Accompanying Documentation
Currently having trouble making an image of the Pi, will upload it as soon as it is available.

In the meantime, please enjoy these two documents. The PDF was created before work began in earnest, and outlined the original vision. The DOC is a research paper describing the process of reaching the current state of Raspberry Fields Forever, as well as its future potential